+++ scripts/release.sh
@@ -9,7 +9,7 @@ read -p "Creating new release for $TAG. Do you want to continue? [Y/n] " prompt
 if [[ $prompt == "y" || $prompt == "Y" || $prompt == "yes" || $prompt == "Yes" ]]; then
     python scripts/prepare_changelog.py
     git add -A
-    git commit -m "Bump version to $TAG for release" || true && git push
+    git commit -m "chore: bump version to $TAG for release" || true && git push
     echo "Creating new git tag $TAG"
     git tag "$TAG" -m "$TAG"
     git push --tags