+++ CHANGELOG.md
@@ -7,16 +7,87 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- A bug when `disable_frontmatter` is ignored for `ObsidianToday` and `ObsidianYesterday`.
+
+## [v1.10.0](https://github.com/epwalsh/obsidian.nvim/releases/tag/v1.10.0) - 2023-05-11
+
 ### Added
 
+- Report errors finding vault from `:ObsidianCheckHealth`.
+- Added `finder` option for choosing a preferred finder backend.
+
+### Fixed
+
+- Removed annoying "skipped updating frontmatter" message on file write.
+
+## [v1.9.0](https://github.com/epwalsh/obsidian.nvim/releases/tag/v1.9.0) - 2023-04-22
+
+### Added
+
+- Added `open_app_foreground` option to open Obsidian.app in foreground on macOS.
+- Added `:ObsidianTemplate` to insert a template, configurable using a `templates` table passed to `setup()`.
+- Added support for following links in markdown format
+- Added `follow_url_func` option to customize behaviour of following URLs
+
+### Changed
+
+- Use `vim.notify` to echo log messages now.
+
+### Fixed
+
+- Gracefully handle invalid aliases / tags in frontmatter (i.e. values that aren't strings). We'll warn about them and ignore the invalid values.
+- Fixed `nvim-cmp` completion for notes that have no `aliases` specified.
+- `nvim-cmp` completion will search based on file names now too, not just contents.
+- Fixed bug when `nvim-cmp` is not installed.
+- Workaround error which prevented users from using `ObsidianOpen` when vault path was configured behind a link
+- Detect URLs when following links and ignore them by default.
+
+## [v1.8.0](https://github.com/epwalsh/obsidian.nvim/releases/tag/v1.8.0) - 2023-02-16
+
+### Changed
+
+- [`lua-yaml`](https://github.com/exosite/lua-yaml) no-longer bundled as a git submodule. Code from that project has been copied and modified into it's own Lua submodule of `obsidian`.
+- (BREAKING) 'nvim-lua/plenary.nvim' is no-longer bundled, so must be explicitly installed (e.g. Plug 'nvim-lua/plenary.nvim' in your `init.nvim`).
+
+### Fixed
+
+- Fixed a bug where creating a new note with `nvim-cmp` completion
+  would cause `nvim-cmp` to stop working.
+- Fixed bug where `disable_frontmatter` setting would be ignored for `:ObsidianNew` command.
+
+## [v1.7.0](https://github.com/epwalsh/obsidian.nvim/releases/tag/v1.7.0) - 2023-02-02
+
+### Added
+
+- Added support for [fzf-lua](https://github.com/ibhagwan/fzf-lua) as one of the possible fallbacks for the `:ObsidianQuickSwitch` command.
+- Added `:ObsidianQuickSwitch` to fuzzy-find a note by name in telescope/fzf _a la_ <C-O> in Obsidian.
+- Added support for [fzf-lua](https://github.com/ibhagwan/fzf-lua) as one of the possible fallbacks for the `:ObsidianSearch` command.
 - Added `:ObsidianFollowLink` and companion function `util.cursor_on_markdown_link`
 - Added `:ObsidianLink` and `:ObsidianLinkNew` commands.
 - Added configuration option `disable_frontmatter` for frontmatter disabling
+- Added line-navigation to `:ObsidianOpen` via the Obsidian Advanced URI plugin
+- Added `:ObsidianYesterday` command to open/create the previous working day daily note
 
 ### Fixed
 
 - Fixed bug with `Note.from_lines` where the given path would be modified in place to be relative to the root, which caused a bug in `:ObsidianFollowLink`.
 - Completion for creating new notes via nvim-cmp is now aware of daily notes, so when you start typing todays date in the form of YYYY-MM-DD, you get a "Create new" completion option for today's daily note if it doesn't exist yet.
+- Fixed bug where `:ObsidianOpen` blocked the NeoVim UI on Linux.
+- `:ObsidianOpen` should now work on Windows.
+- Fixed URL encoding of space characters for better compatibility with external applications.
+- Made more robust to unexpected types in frontmatter.
+- Fixed edge case where frontmatter consisting of exactly one empty field would raise an exception.
+- Fixed `:ObsidianFollowLink` not creating a new note when following a dangling link; matches behavior in the official Obsidian app.
+- Fixed handling spaces in configured vault directory.
+- Fixed `:ObsidianFollowLink` not considering the vault's root directory.
+- Fixed bug where the note ID in the YAML frontmatter wasn't updated after the file is renamed.
+- Fixed `require` module name syntax; see #93 for explanation.
+
+### Changed
+
+- The new note completion source will now create the new note in the same directory as the current note, regardless of the `notes_subdir` setting.
 
 ## [v1.6.1](https://github.com/epwalsh/obsidian.nvim/releases/tag/v1.6.1) - 2022-10-17